+++ src/basic/main.basic.js
@@ -10,14 +10,16 @@ import {
   FLASH_SALE_DISCOUNT,
   SUGGESTION_INTERVAL,
   SUGGESTION_DISCOUNT,
-  BULK_PURCHASE_THRESHOLD,
-  BULK_DISCOUNT_RATE,
-  TUESDAY_DISCOUNT_RATE,
   BONUS_POINT_RATE,
   LOW_STOCK_THRESHOLD,
   MAX_QUANTITY,
 } from './shared/constants.js';
 import { getDOMElements } from './shared/domSelectors.js';
+import {
+  getDiscount,
+  applyTuesdayDiscount,
+  getDiscountRate,
+} from './services/discountService.js';
 
 // store/cartStore.js
 
@@ -25,35 +27,6 @@ let lastSelectedProductId;
 let bonusPoints = 0;
 
 // services/discountService.js
-const getDiscount = (product, quantity) => {
-  if (quantity >= BULK_PURCHASE_THRESHOLD) {
-    return product.discountRate;
-  }
-  return 0;
-};
-
-const applyTuesdayDiscount = (totalAmount, discountRate) => {
-  const isTuesday = new Date().getDay() === 2;
-  if (isTuesday) {
-    const tuesdayDiscountAmount = totalAmount * TUESDAY_DISCOUNT_RATE;
-    const currentDiscountAmount = totalAmount * discountRate;
-    if (tuesdayDiscountAmount > currentDiscountAmount) {
-      return TUESDAY_DISCOUNT_RATE;
-    }
-  }
-  return discountRate;
-};
-
-const getDiscountRate = (productCount, subtotal, totalAmount) => {
-  if (productCount >= 30) {
-    const bulkDiscount = subtotal * BULK_DISCOUNT_RATE;
-    const productDiscount = subtotal - totalAmount;
-    if (bulkDiscount > productDiscount) {
-      return BULK_DISCOUNT_RATE;
-    }
-  }
-  return (subtotal - totalAmount) / subtotal;
-};
 
 // services/cartService.js
