+++ src/basic/services/discountService.js
@@ -1,0 +1,37 @@
+import {
+  BULK_PURCHASE_THRESHOLD,
+  BULK_DISCOUNT_RATE,
+  TUESDAY_DISCOUNT_RATE,
+} from '../shared/constants.js';
+
+const getDiscount = (product, quantity) => {
+  if (quantity >= BULK_PURCHASE_THRESHOLD) {
+    return product.discountRate;
+  }
+  return 0;
+};
+
+const applyTuesdayDiscount = (totalAmount, discountRate) => {
+  const isTuesday = new Date().getDay() === 2;
+  if (isTuesday) {
+    const tuesdayDiscountAmount = totalAmount * TUESDAY_DISCOUNT_RATE;
+    const currentDiscountAmount = totalAmount * discountRate;
+    if (tuesdayDiscountAmount > currentDiscountAmount) {
+      return TUESDAY_DISCOUNT_RATE;
+    }
+  }
+  return discountRate;
+};
+
+const getDiscountRate = (productCount, subtotal, totalAmount) => {
+  if (productCount >= 30) {
+    const bulkDiscount = subtotal * BULK_DISCOUNT_RATE;
+    const productDiscount = subtotal - totalAmount;
+    if (bulkDiscount > productDiscount) {
+      return BULK_DISCOUNT_RATE;
+    }
+  }
+  return (subtotal - totalAmount) / subtotal;
+};
+
+export { getDiscount, applyTuesdayDiscount, getDiscountRate };
